refactor: update import paths for service and provider classes in info and reader services
